requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.5
python-dotenv==0.21.0
redis==4.5.1
gunicorn==20.1.0